Track last activity of mining addresses
Added new column 'last_mine' to 'balance' table that represents the last
time that blockManager added coin to a given address' balance.  This
allows the pool operator to get a general feel for the "activeness" of
its users.

It is MANDATORY to use the ALTER TABLE line in SQL_MIGRATIONS.md dated
7/6/2018 for existing pools, otherwise blockManager will encounter an
error when distributing balance updates.