Fix "Duplicate Job" problem & miner count issue
- The main/worker pool processes were sometimes experiencing a race
condition during template updates due to submitBlock also causing its
own template update check to occur.  The only recurring template update
check that needs to exist is the main pool process, as it triggers an
update to the workers as needed.  This check already happens every
3/10ths of a second.

- When IPs were being banned due to bad shares, they were not being
removed from the miner count for that pool/port.  This is now fixed.