Fix payment error when miner owed payouts from 2+ pool types
A bug in the payment system existed that caused payouts to not be
processed correctly if a regular wallet address was owed balances from
2+ pool types (fees, solo, pplns, and/or pps).  This did not affect
miners using an integrated address.

The bug originates in the payeeObjects object where it would get
overwritten by the multiple balance rows of the same wallet address.

I would eventually like to overhaul the entire payment logic, but
separating the payment run into 4 separate loops (one per pool_type)
solves this for now.